添加数据源日志以及修改SQLite主键类型

+++ src/main/java/cn/schoolwow/quickdao/builder/table/SQLiteTableBuilder.java
@@ -13,7 +13,7 @@ import java.util.List;
 public class SQLiteTableBuilder extends AbstractTableBuilder{
     public SQLiteTableBuilder(QuickDAOConfig quickDAOConfig) {
         super(quickDAOConfig);
-        fieldMapping.put("long", "INT8");
+        fieldMapping.put("long", "INTEGER");
     }
 
     @Override
